Separate type-safe property classes
