+++ common/src/main/java/com/nearsoft/fgaribay/mgmt/config/properties/ExchangeProperties.java
@@ -1,0 +1,34 @@
+package com.nearsoft.fgaribay.mgmt.config.properties;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.validation.annotation.Validated;
+
+import javax.validation.constraints.NotNull;
+
+@Validated
+@Configuration
+@ConfigurationProperties(prefix = "product.exchanges")
+public class ExchangeProperties {
+
+    @NotNull
+    private String creationName;
+
+    @NotNull private String listName;
+
+    public String getCreationName() {
+        return creationName;
+    }
+
+    public void setCreationName(String creationName) {
+        this.creationName = creationName;
+    }
+
+    public String getListName() {
+        return listName;
+    }
+
+    public void setListName(String listName) {
+        this.listName = listName;
+    }
+}
